Add missing function _find_property to datamodels.properties
The method __hasattr__, defined on property nodes, calls
_find_property, which AFAIK was never coded. I was able to determine
that it should test if a property is in the schema and has a default.
So I coded a short function that does that.

I should note that AFAIK there is no double underscore method for
hasattr. It actually calls __getattr__ and tests for an
AttributeError. But since __hasattr__ is part of the interface and
apparently is being used, I coded _find_property.