Modify schemas where ndim is defined for dq but not data
The ndim field in a schema is used to validate the number of
dimensions and to generate a default array when an fits hdu is
undefined. Setting ndim on the dq property but not the data property
causes an error when a datamodel is copied. This fix moves the ndim
field in the schema from the dq property to the data property.