Modify schemas where ndim is defined for dq but not data (#1135)
This fixes a problem with datamodels copy reported by @philhodge 

The code which writes EXTVER is controlled by the sequence_index field
in the schema. The default value (in the code, not the schema) is
None, which means EXTVER is not written if squence_index is not in the
schema. Setting the default to zero causes EXTVER to be set to one when
sequence_index is not in the schema.

* Move wcsinfo to science extension

The core schema has been updated to read and write the wcs from the
science extension instead of the primary header. Tests and test data
have been updated to match. Default arguments of get_fits_wcs and
set_fits_wcs have been updated. A hdu_ver argument has been added to
get_fits_wcs.

* Add fits_required to schema verification

The new field fits_required can be added to a schema. When it is
present, the data named in the schema must be present in the fits
file. The data can either be a keyword or extension. Normally the code
will throw a validation error if the data is missing, but if
pass_invalid_value is true, it will give a warning instead.

* Add missing function _find_property to datamodels.properties

The method __hasattr__, defined on property nodes, calls
_find_property, which AFAIK was never coded. I was able to determine
that it should test if a property is in the schema and has a default.
So I coded a short function that does that.

I should note that AFAIK there is no double underscore method for
hasattr. It actually calls __getattr__ and tests for an
AttributeError. But since __hasattr__ is part of the interface and
apparently is being used, I coded _find_property.

* Modify schemas where ndim is defined for dq but not data

The ndim field in a schema is used to validate the number of
dimensions and to generate a default array when an fits hdu is
undefined. Setting ndim on the dq property but not the data property
causes an error when a datamodel is copied. This fix moves the ndim
field in the schema from the dq property to the data property.