Fix hdulist warning message caused by double close
Under some circumstances, users would get a warning message when
closing a model file that said the hdulist was already closed. The
warning has been traced to the list of files to close contained in the
model. When a model is copied, the same file is on the list of both
models, so the same file gets closed twice. The code has been changed
to truncate the list of files to close when a model is copied, so the
file is only on a single list.

This problem was first reported by @PhilHodge.