Place wcsinfo metadata in its own schema
The wcsinfo metadata is now read from the SCI extension. However, some
image types do not contain a SCI extension and others contain multiple
SCI extensions. For these images wcsinfo is not relevant, so wcsinfo
should not be in the core schema. It was moved out into its own schema
and those schemas with a single SCI extension were modified to include
it. One test needed to be modified so that it creates an ImageModel,
which has wcsinfo, instead of a DataModel, which does not.