Pull additional keywords out of core schema and update multi* schemas
Any keyword in the core schema that points to the SCI extension has
been moved to the wcsinfo schema. Schemas which represent multiple
sets of obsevations, multiexposure, multiproduct, and multislit have
been updated to include the wcsinfo schema. The MultiExposureModel has
code that's tightly tied to the layout of its schema. The code has
been updated to work with the new schema.