There's no longer a need for two get_subarray methods.
Method get_subarray was replaced by get_subarray_trapdensity and
get_subarray_persat because model.copy() didn't work for these files
(see pull request #1135).  Because that has now been fixed, the change
for this commit is to return to just one get_subarray method, but to
retain the fixes for the other bug reported in issue #1129.