Add attributes slit_ra and slit_dec to MultiSpecModel.
The right ascension and declination at the center of the extraction
region are now assigned to these new attributes.  These are the same
values as are returned by the wcs object, which is possible because
the coordinates are independent of row number in the table of extracted
values.  See issue #1034.

Additional metadata attributes have also been added to the MultiSpecModel.
These are based on attributes that were added not very long ago to the
MultiSlitModel, but not all are included in MultiSpecModel.  The extract_1d
step does not actually populate these yet.  See issue #1042.