Slit- and source-related keywords are copied to output.
Attributes such as the slit name, source name, and source location are now
copied from the input file (if it's a MultiSlitModel or MultiProductModel)
to the output MultiSpecModel file.  For other models (except CubeModel),
the slit name will be copied to the output, but other attributes are not
known or are not relevant, so they will not be assigned.  See issue #1042.