Change creation date code
Datamodels adds the creation date of a file into the metadata when
saving it. The current code adds the date to the metadata and then
updates the format. This assumes that saving to the metadata does not
change the format. It's safer to update the format of the date and
then add it to the metadata.