Modify registry initialization so it is only invoked oned
The __init__ method for datamodels can be call more than once. But
this astropy.io.registry code throws an error if it is. So code was
added to check if the __init__ method has been called previously and
only do the registry initialization if it has not been called
previously.