Update extract_1d to handle ModelContainer as input
If the input to the extract_1d step is a ModelContainer, extract_1d_step.py
now loops over the objects in the container and creates one MultiSpecModel
for each input.  If there is only one object in the input, the output will
be a MultiSpecModel; however, if the input contains more than one object,
the output will be a ModelContainer with one MultiSpecModel for each input
object.

If the input to the extract_1d step is not a ModelContainer, the step
will write a single MultiSpecModel as it did prior to these changes.
See issue #1264.