Support background subtraction for WFSS/GRISM data
The code that actually does the background subtraction for WFSS/GRISM data
cannot be executed yet because we don't have the required reference files.
The line that defines reference_file_types is commented out, and the lines
that open the reference file and assign the model to the list of reference
file models are also commented out.  See issue #599.