Remove extra_fits when readung and writing model file from disk
The extra_fits area in a datamodel is intended to hold the information
from a file that is not contained in the model schema. It is
tempoarary and should not be persisted to disk in the asdf extension
when saving a model file. This change removes the extra_fits section
of a tree after loading the asdf extension from a file and before
writing the asdf extension to a file. In addition, it checks to make
sure that the asdf extension is not saved in the extra_fits area.