Fix for jump detection when there are saturated groups
The program was using a wrong index to find the largest non-saturated difference.
Also, the calcuation of the median was not correct since it assumed all pixels in the integration had the same number of non-saturated groups.
An expcected to fail test was added for a future feature to flag pixels adjacent to detected jumps.