Check whether nrs_normal and nrs_reference is None
The code had been using try/except AttributeError to check for missing
keywords NRS_NORM or NRS_REF, but now the relevant attributes are present,
and a missing keyword just results in a None value for the attribute, so
the code now checks for "is None".
The return statements in `process` had been rearranged, and a new return
needed to be added to the section for IRS2 data.
See issue #1279.