Update to support WFSS/GRISM data.
Expanded allowed extraction limits to go from the lower edge of the first
pixel to the upper edge of the last pixel (a half-pixel difference at each
end).

Added a test to skip exposures with a zero-length axis.

Added code to get the spectral order number, and to include it (if necessary)
in the call to the wcs function and save it in the output file.  For GRISM
data, loop over pixels rather than passing arrays when calling wcs().

MultiSpecModel schema was modified to include spectral_order.

See issue #1297.