Skip NaNs when picking RA and Dec for keywords
The `extract` method in `ExtractModel` has been modified to check for
NaNs in the right ascension and declination returned by the WCS function
because in some cases (this was seen for slit S1600A1) there are NaNs.  The
right ascension and declination are assigned to keywords SLIT_RA and SLIT_DEC,
and NaNs are not valid for keyword values.  See issue #1356.