Update schema editor to support adding new schema fields
The schema editor compares the model schemas in the datamodels
directory to the keyword databse files and either reports the
differences (if list=True) or edits the model schemas to match the
keyword database (if list=False.) Previously the schema editor did not
add fields in the keyword database to the model schemas, as there was
no information about which schema to add the fields to. But the
decision was made to all keywords to the core schema, so now the
schema editor does that. The code to delete model schem fields, edit
them, and rename them is bascially the same as before. However, when
modifying the existing code some bugs were found and fixed and a
number of changes to streamline the existing code were made as
well. So the code is considerably different from the prvious version.