Revert order of writing model data
The order of writing model data has been reverted so that schema data
is written before extra_fits data. The order was previously switched
so that extra_fits was written first. This was to work around a
problem where the same extension would be written both from the schema
and extra_fits, resulting in two copies of the extension in the
file. That problem has been fixed in this change by having
_save_extra_fits check to see if the extension already exists and not
creating an extension if it does.