Unset model_type when opening images as DataModel and with open
If the DATAMODL keyword is not found in the primary image header,
DataModel and its inheritors add it. If an image is opened with the
open() function, and DATAMODL is not set, open() should not set it,
because the class it chooses is only a best guess. So this update sets
model_type to None after opening the image, so DATAMODL is not written
to the header. In addition, when an image is opened with the base
class, DataModel, DATAMODL should not be set, as there is no image of
the base type. So this update also sets model_type to None in that
case. This update also moves all the code that sets model_type to
__init__ from on_save, so that code which calls DataModel can override
model_type.