Get hdu type from value when saving extra_fits data
Data stored in the extra_fits area can either be an image or a
bintable. Previously the code assumed it was an image. This update
pulls some code that was in _make_new_hdu into a new function,
_get_hdu_type, and calls it from _make_new_hdu and _make_hdu. This
ensures that whenever a new hdu is written, the code first checks the
type of the data (if present) and selects the proper class to create
the new hdu.