Two refinements to CRDS client optimized glue code:
1. Cached dataset headers are constrained to simple Python values to preclude
complex memory leaks.

2. Exception-based header and results caching is modified so that the uncached
calls do not occur within the "not already cached" exception.  This prevents
confusing error messages where the original exception is driven simply because
the cache is empty but the secondary exception occurs while trying to do real
computes and fill the cache.  In the case of the outer exception,  a result may be
computable... it just isn't in the cache.  By unnesting the exceptions appropriate
error handling is restored.