Fix group_id logic and related tests
Due to #795, got a test failure due to both bad logic in the change
and adding the logic invalidated the tests. The bad logic was that if
a model already has a `group_id`, don't create another one, which is
fine. However, the grouping code was then bypassed completely.

With fixed logic, the tests' logic was no longer valid. The tests
depended on `group_id` always being dynaically created. Added a reset
function allowing the tests to perform as expected.