Remove filename from the list of protected keywords (#1531)
Previously three "protected keywords" were not copied by the update
method. The three is now two: filename is copied and date and
model_type are not. While not copying filename was correct, current
use of the update method by the jwst pipeline turned a feature into a
bug. (Or a bug into a feature, depending on which way you look at it.)