Check whether ramp_fit was skipped
The ramp_fit step normally returns two objects.  When ramp_fit is run
in calwebb_detector1, however, the user can specify that the step be
skipped, in which case the step will return just one object, the input
model.  This will fail in the pipeline because the code explicitly
receives two objects from the step, but only one was returned.  This
can be (temporarily) fixed by checking the `skip` flag.  See issue #1532.