Use wavelength array in preference to wcs function.
For NIRSpec spectroscopic data, the flat_field step computes the wavelength
at each pixel by calling the WCS function (except for BRIGHTOBJ data, which
already get the wavelength from the wavelength attribute).  If the wavelength
attribute exists and is populated (i.e. not all zeros), that should be used
for the wavelengths instead.  See issue #1374.