Correct potential problem with indexing in flat_field
When looping over input slits for NIRSpec data, and for the case that an
output file with the computed flat fields will be written, set the flat field
to a constant 1 and data quality to 0 for any input slit for which processing
is skipped.  Also set the output wavelength attribute to constant 0.
See issue #1547.