Remove tables from asdf tree before serializing it
When an datamodels object is saved to a fits file the asdf tree
representing it is serialized to a fits extension. If the original
file has a bintable extension, the asdf extension also includes it
inline. As a result there are two copies of each binary table in the
saved fits file. This does not happen for image extensions, they are
not serialized. While this is a problem in the asdf code, for now I
have added a patch to fits_support.py. A new function, _snip_tables,
removes tables from the asdf tree before it is serialized. This fix
will be removed after the problem in the asdf code is identified and
fixed.