Fix default value for sovler parameter to match_lsq
The default given for parameter `solver` was `LU`.
However, according to comments and code, only `RLU` and `PINV`
are supported.

Though it seems to be just a typo and the intent was for `RLU`
to be default, I went with `PINV` because then the code would
actually run for MRS test data.