Update combine_1d
Changes include:
The FLUX, ERROR, and DQ columns are combined, in addition to the NET.
The input DQ column is checked for pixels flagged with DO_NOT_USE, and
such pixels will not contribute to the output.
Output columns (except for DQ) will be double precision.
An output WCS model will be created using code in extract_1d.
An open output model is returned, rather than explicitly writing an
output file.
See issue #1551.