Allow 1-D spectral extraction for multiple spectral orders.
In order to support the option to extract more than one spectral order,
the reference file can have a "spectral_order" key in addition to the "id"
key.  While "id" is still the primary key to be matched for selecting
the relevant information, "spectral_order" can now be used as a secondary
selection criterion.  If there is an aperture in the reference file for
which the "id" key matched, that's considered to be (at least) a partial
match.  If there is no "spectral_order" key for this "id", or if
"spectral_order" also matches, that's regarded as an exact match.  A
spectrum will be extracted only if the match is exact.  If the match is
partial, the specified (but not found) spectral order will be skipped; if
there's not even a partial match, that's an error.
A new exception has been defined, InvalidSpectralOrderNumberError, which
ExtractModel will raise if the spectral order number is invalid.  At present,
this only applies to NIRISS SOSS data.  This distinguishes between an
invalid spectral order that is nevertheless included in the reference
file, and a spectral order that is not found in the reference file.

Function load_ref_file was added.  This opens the JSON reference file and
reads it into a dictionary, which is returned.  Function
get_extract_parameters was modified to accept that dictionary, instead of
reading and calling json.load for every slit or spectral order.

A `verbose` argument was added to extract_one_slit, to suppress logging of
extraction parameters, and the log.info call for NIRISS SOSS spectral order
has been deleted, since that value is printed elsewhere.  The need for this
argument became apparent when testing with CubeModel data, in which case
the logging of (identical) extraction parameters was repeated for every
integration.

No change was needed to support following a spectral trace that is
defined by polynomial functions for the lower and upper limits.  That
option was added by Mihai Cara more than a year ago.

See issue #1626.