Fix index out of bounds errors.
Changes were made to function _extract_colpix in extract1d.py to handle
the possibility that both limits of an extraction region are above (or to
the right of, depending on dispersion direction) the input image.  The
number of pixels in the extraction region can be zero, and arrays were
allocated with that length; then a value was assigned to element 0.  The
other case was where there was a check that the lower limit was greater
than -0.5 (the lower edge of the input image), but there was no check that
the lower limit was strictly less than the upper edge of the input image;
the index for that case was one element outside the image.

See issue #1673.