Remove code to remove fits tables from asdf tree
A bug in asdf caused it to save a serialized version of fits tables in
the asdf extension, increasing the size of the fits file with
duplicate code. I added a short function to fits_support to remove
tables from the serialization. Now that the bug in asdf has been
fixed, it's no longer needed, so I have removed it.