Only call the function for saturated data if some groups are saturated.
The function `predict_saturation_capture` for predicting the number of
charges trapped during the saturated portion of a ramp is now called
with a mask to include only pixels for which at least part of the ramp is
saturated (in the sense of being greater than the persistence saturation
limit).  See issue #1599.