Use DQ extensions of persat and trap_density
The slope is set to zero for pixels in the persistence saturation limit
reference file that are flagged with DO_NOT_USE.  The trap density is set
to zero for pixels in the trap density reference file that are flagged
with DO_NOT_USE (resulting in no increase in the number of filled traps).
See issue #1635.