Convert numpy records to fits records when setting properties
It was noted that when creating new spectral tables with datamodels
the code could not subsequently set the table column characteristics,
such as units. This is because numpy records do not support column
definitions. In order to fix this prblem a new function, _as_fitsrec
was added to datamodels.properties. It converts numpy records into
FITS_rec objects and adds column definitions derived from tha numpy
record. This change modified several test results in test_schema.py,
which were updated to agree with the new code.