Replace NaNs and DO_NOT_USE in science data with 0.
Function replace_bad_values was added.  Pixels in the science data that
are NaN or are flagged in the dq attribute (if that exists) with DO_NOT_USE
will be set to 0.  The changes are made to a local copy, not to the data
in the input model.  See issue #1788.