NIRSpec WCS input coordinates are 1-based (flat_field)
For data taken with instruments other than NIRSpec, the wavelengths are
not used and the wcs function is not called, so no change was needed
for those data.

For NIRSpec data, changes were made to three functions:
(1) NIRSpec_brightobj for NRS_BRIGHTOBJ data; (2) NIRSpec_IFU for NRS_IFU
data; and (3) do_NIRSpec_flat_field for all other NIRSpec data.  The change
was to increment the pixel coordinates by 1 before calling the wcs function.

See issue #1781.