Updated gain_scale_step.py and several reference file schema where GAINFACT (meta.exposure.gain_factor) is deemed relevant enough to warrant a CRDS warning when absent... and hence needs data models support so developers can add it.
For consistency with the core schema,  I moved the reference file
mapping for GAINFACT from meta.gain_factor to meta.exposure.gain_factor.
This required a corresponding update to the gain_scale_step module.

As an aside, inconsistency between the core schema and reference
file schemas is currently a technical issue for CRDS,  CRDS essentially
needs something like datamodels_to_fits() to be 1:1 and invertible
for all schema:

meta.exposure.gain_factor  -> GAINFACT
GAINFACT -> meta.exposure.gain_factor

This enables CRDS to use the core schema to define the relationship
for all types.   It's also good for average developers who believe
datasets and references line up in an apples-to-apples relationship.