Truncate arrays where wavelength is NaN.
Function nans_at_endpoints was added.  This flags NaNs in the wavelength
array with DO_NOT_USE in the dq array.  If there are NaNs at endpoints
of the wavelength array, all the output arrays will be truncated to remove
the contiguous NaN elements at either or both endpoints.  If there are any
interior points in the wavelength array that are NaN, these will be left
as NaN, though still flagged as bad in the dq array.  See issue #1803.