Add make_header script
The make_header script creates an image containing all the required
header keywords for a pipeline. The pipeline is selected based on the
first three command line arguments: instrument name, mode (imaging or
spectral) and pieline level (1, 2, or 3). The 4th command line
argument is the name of the output file produced.