Allow conversion between model types in datamodels
Datamodels has been modified to allow conversion between model
types. The method is to open a datamodel of the desired type using
a datamodel of another type as its argument. For example,

    im = datamodels.open('reference_file.fits')
    im = FlatModel(im)

The first line opens the image as a generic reference file. The second
line converts it to a flat field reference file.