Allow conversion between model types in datamodels (#1857)
Datamodels has been modified to allow conversion between model
types. The method is to open a datamodel of the desired type using
a datamodel of another type as its argument. For example,

    im = datamodels.open('reference_file.fits')
    im = FlatModel(im)

The first line opens the image as a generic reference file. The second
line converts it to a flat field reference file.

Two tests needed to be modified to support casting because previously
casting would raise an error and the tests checked for this behavior.

I placed the validation performed after casting into its own method so
it be called by other methods that alter the metadata or schema.