Clean up datamodels model_base __init__
The __init__ function of model_base is an ungainly beast that I've
tried to tame. The current code is correct, but I've re-arranged it
and added comments. There's one improvement: intializing a datamodel
from a shape fails if the model has no data section, as it should. I
catch the confusing attribute error and raise a more explanatory type
error instead.