Update make_header to address testing issues
We asked for feedback on the make_header script. Problems were reported
in issue #1780. The header file has both too few and too many
keywords. Some keywords used to select reference files were
missing. This was addressed by modifying additional_metadata to
incluse all the instrument metadata used to select reference files. It
also included instrument specific keywords irrelevant to the
instrument and mode passed to make_header. This was addressed by
checking a new function, another_instrument, which looks at the title
in the schema to see if a metadata item is associated with an
instrument or exposure type.