Add read_schema function to datamodels (#1892)
When creating a DataModel you have the option of passing it a schema
instead of using the predefined schema in the datamodels schema
subdirectory. This si useful when debugging or testing schema
changes. However, since it is a schema instead of the name of a file
containing a schema, it is difficult to access this functionality. The
read_schema function has been added to schema.py to make the job easier.
Its calling sequence is:

    schema = schema.read_schema(schema_file, extensions=None)

You can then pass the schema as a argument to the datamodel you are
creating. The optional parameter extensions is for passing
non-standard extensions to the schema, just as when creating a
datamodel.