For extract_1d, find nod/dither offset from the WCS
Function locn_from_wcs was added.  For one column in the dispersion
direction (the column in the middle of the WCS bounding box), this function
determines the cross-dispersion location (i.e. the pixel in that column)
of the target spectrum.  (I'll describe the case for dispersion in the
horizontal direction.)  The target right ascension and declination are
gotten from keywords TARG_RA and TARG_DEC respectively.  For each pixel in
the middle column, the right ascension, declination, and wavelength are
obtained via the WCS function.  The cross-dispersion location is the pixel
along that column for which the distance from the target coordinates to the
computed coordinates is closest to zero.  Unless the pixel closest to the
target was at an endpoint of the aperture, the nod/dither offset will be
set to the difference between this pixel and the nominal (i.e. from the
reference file) location of the target spectrum.  Any nod offset amplitude
that was specified in the reference file (e.g. using keywords such as
"nod2_offset", "nod3_offset", and "nod5_offset") will be ignored.  Applying
the shift is now done by method add_nod_correction, regardless of whether
the reference file is JSON format or an image.

Function find_dispaxis was added.  This determines the dispersion direction
by comparing the rate at which wavelength changes from pixel to pixel, in
the horizontal and vertical directions.  If dispaxis was specified in the
reference file, that value takes precedence.  If dispaxis was not specified,
or if there was no reference file, the value found by find_dispaxis will
be used.  This was necessary for the case that there was no reference file,
and for other cases it makes it possible to omit dispaxis from the reference
file.  If there was not sufficient information in the science file to
determine the dispersion direction, a warning will be printed, and no change
to dispaxis will be made.

See issue #1894.