Add __str__ method to DataModels (#1927)
The method returns a string containing the data type, shape, and filename
the datamodel was read from, where these are available. An example of the
result is:

    <MaskModel(2048, 2048) from nircam_mask.fits>

The new method is useful when logging and debugging.

As I just learned, Python does not call __str__ on each item when a
compound object is stringified. It calls __repr__ instead. Since this
is our use case, I have renamed _str__ to __repr__.