Fix history validation error
Datamodels tests were throwing a validation error when writing new
history records. It's not clear why we started seeing the errors,
perhaps an update to asdf. But the cause is that there were two
different data structures to hold the history records and only one of
them passed validation. The solution is to convert the invalid
structure to the valid one before serializing the asdf file to its
extension.