Allow one or two of the NIRSpec flat fields to be missing.
Tests were removed from flat_field_step.py that prohibited the step from
running (for spectroscopic data) if any of the three NIRSpec flat fields was
missing (meaning that the file name was "N/A").  Now, any one or two of the
flat field reference files may be missing.  If no flat field reference file
is found, however, this is logged as a warning and the step will be skipped.

In flat_field.py, the main change was to add a test at the beginning of
each of fore_optics_flat, spectrograph_flat, and detector_flat.  If the
respective f_flat_model, s_flat_model, or d_flat_model is None, the function
creates an array of ones for the f_flat (s_flat, d_flat) and returns
that array, along with None for the dq array.

See also issue #1914.