New script creates world coordinates files from extract2d files
The script is a tool to read in the output of extract2d (FS and MOS)
or assign_wcs (IFU) and apply the WCS transforms to all pixels in a
slit. For each slit it writes the results as a cube with four planes
(wavelength, ra, dec, y_slit) in a separate fits extension.  The file
is saved with a suffix "world_coordinates".