The extract_1d step should copy data from the integration times table
Function populate_time_keywords was added to the extract_1d step.  This
copies the start, middle, and end times for each integration from entries
in the INT_TIMES table in the input file to header keywords in the output
file.

Various schema files were modified to include the INT_TIMES table or the
new keywords in the MultiSpecModel.  core.schema.yaml was also modified
to include the start and end integration numbers for segmented data,
which are needed to find the relevant rows in the INT_TIMES table.

See issue #1881.

The extract_1d step was also updated to treat EXP_TYPE = "NRC_WFSS" the
same as "NRC_GRISM".  See issue #1985.