Add support for coronographic pipeline to make_header
This version of make_header allows you to select coronographic
exposure types and if selected will produce a header with keywords
used by the corongraphic pipline. The changes to support that are in
choose_pipeline.

Testing uncovered a problem with stripping method calls from metadata
names. The new code uses a simpler algorithm. It looks for a
parenthesis following the name. Changes are to the pattern used in
undefined_metadata and to the function drop_suffix.

I also added additional diagnostic prints in the function
log_file. These prints are normally turned off.