For extract_1d, find nod/dither offset from xoffset and yoffset
Finding the nod/dither offset by using the WCS function did not work.  An
alternative is to use keywords XOFFSET and YOFFSET.  These can't be applied
directly, because they are in units of arcseconds in the Ideal coordinate
system, not detector pixels.  Method offset_from_offset was added, which
gets these offsets and converts to offsets in detector coordinates.  Due
to the lack of test data at the present time, however, the computed offset
will not actually be applied to the data.  If the offset is not zero, it
will be written to the log using DEBUG level, so it can be tested when
suitable data become available.

Several calls to log.error were changed to log.warning.

See issue #1894.