Add additional information to validation error message
This change catches metadata validation errors and adds a line at the
start of the error with the name of the attribute being validated. The
line looks like:

    While validating X the following error occurred:

The reason for the additional line is to make it easier to determine
the cause of the error.

This change also addresses an error encountered during testing. The
code which sets a metadata item to its default value now only does so
if the value is not None. It's never correct to set the value to None
because this causes the type of the metadata item to be invalid.