In find_dispaxis, regard wavelength = 0 as invalid
The find_dispaxis function used the wavelengths to determine which is the
dispersion axis.  The wavelengths may be taken from the wavelength attribute
(extension in the input file), or they may be computed by using the wcs
function.  Wavelengths are not necessarily defined over the entire data
range, and the previous assumption that the wavelength would be NaN for
pixels for which the wavelength could not be determined was not always
correct; sometimes the wavelength is 0.  find_dispaxis has been modified
to treat both 0 and NaN as invalid.

Also in find_dispaxis, it was necessary to explicitly set the value of
the spectral order for NIRISS SOSS data.

See issue #2031.