Add support for compressed image extensions to datamodels
This fix allows programs to read compressed image extensions from fits
files and update them. The main change has been to how the code
determines the type of an extension. It no longer assumes that anything
that's not a binary table is a regular image extension. The code that
reads and writes compressed images is layered on top of the astropy fits
library code.