Update make_header script
An independent review of the script in #2048 found several
problems. The script was not treating metadata keywords marked as IRS2
as nirspec specific. Bugs in the code kept it from setting metadata
with enumerated values or correctly choosing the instrument that a
metadata keyword was associated with. Re-examination of the jwst
source code led to some changes to default values used for instrument
and modes. And in some cases, defaults depend on exposure type as well
as instrument.