extract_1d should not repeat log messages for every integration
A verbose argument was added to a number of functions, initially set to True
but reset to False after the first integration of a multi-integration
exposure.  A new parameter log_increment was added to the step.  This lets
the user specify how frequently INFO messages are written during processing
of a multi-integration exposure.

See issue #2079.