Add a custom validator that checks for None
Datamodels allows users to set a metadata keyword to None, which
indicates that it should not be written to the fits header. In certain
cases datamodels was checking a value of None against the type of the
keyword in the schema and throwing a warning. To prevent this from
happening this update adds a custom validator for type which skips
validation when the keyword value is None.

This update also moves all the validation code from util.py into a new
file, validate.py. The amount of validation code in datamodels was
large enough to deserve its own file. This also required updating all
the places where validation was called, including tests.