Do not write asdf extension when saving reference files
The asdf format has had compatibility problems between software
versions that has complicated the handling of reference files. So
datamodels this code change updates datamodels to not write an asdf
extension to reference files. There are three changes:

Datamodels has a new attribute, _no_asdf_extension. It is set to True
for reference files and False for other files. When a datamodel is
saved to a file, the code checks this attribute and calls the fitsio
writeto method if it is True and asdf write_to method if it is False.

All reference file model types are now subclassed from
ReferenceFileModel, which sets _no_asdf_extension to True.

The validate method in ReferenceFileModel now calls the validate
method in the base class of Datamodels so that full validation is done
when updating reference files.