Rewrite datamodels history getter and setter
The history records in datamodels are best represented as a list that
one appends to. This update rewrites the history getter and setter to
implement this by:

Having the getter return a list of history entries. It first creates
the list of entries if it does not exist.

Rewriting the setter so that it appends to the list of history
entries. If the value is not a History Entry, it is converted to a
HistoryEntry.

Due to the different ways values can be added to the history, it is
possible to wind up with a mixed list of HistoryEntry's and other
values. This is fixed when the history is written to the fits header.