Wrap validation asserions with test for strict_validation
Reference file models have a validate method which contain assertions
about its properties. This update wraps the assertions in a try block.
If the exception is thrown, it tests strict_validation. If it is true,
the error is re-thrown. If it is false, the exception is converted to
a warning. This update also adds a call to the super-class validate
method to a few cases where it was not being called by a reference
file model.